% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_et_st}
\alias{sample_et_st}
\title{Sample environment process and state process of a subject}
\usage{
sample_et_st(rt, wb_shape, wb_scale, p, q, e_stat, s_stat, ntrial, nstate)
}
\arguments{
\item{rt}{a vector of response time sequence of a subject}

\item{wb_shape}{current update of Weibull shape parameter}

\item{wb_scale}{a vector of current update of Weibull scale parameters in all states}

\item{p}{an array of current update of transition matrices of state process}

\item{q}{a matrix of current update of transition matrix of environment process}

\item{e_stat}{a vector of current update of stationary distribution of environment process}

\item{s_stat}{a vector of current update of stationary distribution of state process}

\item{ntrial}{total number of trials}

\item{nstate}{total number of states}
}
\value{
A list with the following components:
  \item{et}{a vector of Gibbs sample of environment process}
  \item{st}{a vector of Gibbs sample of state process}
}
\description{
Produce a Gibbs sample of environment process and state process of a subject
}

