% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_log_shape.R
\name{sample_log_shape}
\alias{sample_log_shape}
\title{Sample Weibull log shape parameter}
\usage{
sample_log_shape(rt, st, y, eta, prior_mu, prior_sigma, proposal_std, nstate)
}
\arguments{
\item{rt}{a vector of response time sequence of a subject}

\item{st}{a vector of current update of state process of a subject}

\item{y}{a vector of current values of Weibull log scale parameters in all states}

\item{eta}{current value of Weibull log shape parameter of a subject}

\item{prior_mu}{mean of normal prior distribution of Weibull log shape parameter}

\item{prior_sigma}{standard deviation of normal prior distribution of Weibull
log shape parameter}

\item{proposal_std}{random walk proposal standard deviation}

\item{nstate}{total number of states}
}
\value{
A vector whose first element is a Gibbs sample and second element 
        is a binary indicator value of acceptance or rejection of the proposed 
        sample in random walk Metropolis–Hastings algorithm
}
\description{
Produce a Gibbs sample of Weibull log shape parameter of a subject
}
