% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_p.R
\name{sample_p}
\alias{sample_p}
\title{Sample transition matrices of state process}
\usage{
sample_p(p_all, q, prior_mat, et, st, ntrial, nstate)
}
\arguments{
\item{p_all}{a matrix of current update of transition matrices of state process, 
transition matrices under different environments are stacked vertically}

\item{q}{a matrix of current update of transition matrix of environment process}

\item{prior_mat}{a matrix of concentration parameters of Dirichlet priors,
each row corresponds to a row of transition matrix}

\item{et}{a vector of current update of environment process of a subject}

\item{st}{a vector of current update of state process of a subject}

\item{ntrial}{total number of trials}

\item{nstate}{total number of states}
}
\value{
A list with the following components:
 \item{p}{a Gibbs sample of transition matrices of state process}
 \item{accept}{a binary indicator vector of acceptance or rejection of proposed 
               sample of each row of transition matrices in the independent
               Metropolis–Hastings algorithm}
}
\description{
Produce a Gibbs sample of transition matrices of state process under all environments of a subject
}

