% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_q.R
\name{sample_q}
\alias{sample_q}
\title{Sample transition matrix of environment process}
\usage{
sample_q(q, et, prior, ntrial, nenv)
}
\arguments{
\item{q}{a matrix of current update of transition matrix of environment process}

\item{et}{a vector of current update of environment process of a subject}

\item{prior}{a matrix of concentration parameters of Dirichlet priors,
each row corresponds to a row of transition matrix}

\item{ntrial}{total number of trials}

\item{nenv}{total number of environments}
}
\value{
A list with the following components:
 \item{q}{a Gibbs sample of transition matrix of environment process}
 \item{accept}{a binary indicator vector of acceptance or rejection 
               of proposed sample of each row of transition matrix 
               in the independent Metropolis–Hastings algorithm}
}
\description{
Produce a Gibbs sample of transition matrix of environment process of a subject
}

