% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_ref_std.R
\name{sample_ref_std}
\alias{sample_ref_std}
\title{Sample the standard deviation of subject random effects}
\usage{
sample_ref_std(alpha, ref_std, s, nsubj, prior_hc_scale, prop_std)
}
\arguments{
\item{alpha}{a matrix of current update of subject random effects}

\item{ref_std}{current value of the standard deviations of subject random effects}

\item{s}{a specific state}

\item{nsubj}{total number of subjects}

\item{prior_hc_scale}{scale parameter of half-Cauchy prior}

\item{prop_std}{random walk proposal standard deviation}
}
\value{
A vector whose first element is a Gibbs sample and second element 
        is a binary indicator value of acceptance or rejection of the proposed 
        sample in random walk Metropolis–Hastings algorithm
}
\description{
Produce a Gibbs sample of the standard deviation of subject random effects in a state
}
