% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_subj_ref.R
\name{sample_subj_ref}
\alias{sample_subj_ref}
\title{Sample subject random effects}
\usage{
sample_subj_ref(y, mu, var_logscale, ref_std, i, s)
}
\arguments{
\item{y}{a matrix of current update of Weibull log scale parameters}

\item{mu}{a vector of current update of intercept effects in all states}

\item{var_logscale}{a vector of current update of variances of Weibull log scale parameters in all states}

\item{ref_std}{a vector of current update of standard deviations of subject random effects}

\item{i}{a specific subject}

\item{s}{a specific state}
}
\value{
A Gibbs sample of subject random effect of a subject in a state
}
\description{
Produce a Gibbs sample of subject random effect of a subject in a state
}
