% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stat_dist}
\alias{stat_dist}
\title{Compute stationary distributions of environment process and state process}
\usage{
stat_dist(p, q, nstate, nenv)
}
\arguments{
\item{p}{an array of transition matrices of state process}

\item{q}{a matrix of transition matrix of environment process}

\item{nstate}{total number of states}

\item{nenv}{total number of environments}
}
\value{
A list with the following components:
  \item{e_stat}{stationary distribution of environment process}
  \item{s_stat}{stationary distribution of state process}
}
\description{
Compute stationary distributions of environment process and state process
}
