% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_intercept.R
\name{sample_intercept}
\alias{sample_intercept}
\title{Sample intercept effects}
\usage{
sample_intercept(y, alpha, var_logscale, m_mu, sigma_mu_inv, nsubj)
}
\arguments{
\item{y}{a matrix of current update of Weibull log scale parameters}

\item{alpha}{a matrix of current update of subject random effects}

\item{var_logscale}{a vector of current update of variances of Weibull log scale parameters}

\item{m_mu}{a vector of means of multivariate normal prior of intercept effects}

\item{sigma_mu_inv}{inverse covariance matrix of multivariate normal prior 
of intercept effects}

\item{nsubj}{total number of subjects}
}
\value{
A Gibbs sample of intercept effects of all states
}
\description{
Produce a Gibbs sample of intercept effects of all states
}

