% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_std_logscale.R
\name{sample_std_logscale}
\alias{sample_std_logscale}
\title{Sample the variance of Weibull log scale parameter in a state}
\usage{
sample_std_logscale(log_scale, mean_log_scale, std, nsubj, prior_hc_scale,
  prop_std, s)
}
\arguments{
\item{log_scale}{a vector of log Weibull scale parameters across all 
subjects in a state}

\item{mean_log_scale}{a vector of means of log Weibull scale parameters
across all subjects in a state}

\item{nsubj}{total number of subjects}

\item{prior_invgamma_shape}{shape parameter of the inverse-gamma prior}

\item{prior_invgamma_scale}{scale parameter of the inverse-gamma prior}
}
\value{
A sample from the full conditional of the variance of Weibull log 
        scale parameter in a state
}
\description{
Produce a Gibbs sample of the variance of Weibull log scale parameter in a state
}

